Add (parent dir of) current working dir to python search path
Try to find the lowest directory, which does not contain a __init__.py and use
that. That way starting in a subdirectory will also find the app when importing
the MARA_APP module and when searching for local_setup.py.

Influenced by how flask does it.

The alternative is to actually install the app.