Add config init helper to be used in scripts
A lot of scripts need the config available, but won't be implmented as mara click commands.

For these there is now a small helper which calls all the app composing steps so that the config system is in the same state as it would be if called as a mara command.