Find MARA_APP_ROOT and add it to sys.path
This is needed to get python mapper scripts and so on running.

There was an interesting fact that if I tried importing app.app in  a
directory which contained a __init__.py, the later calls would also fail. I
couldn't reproduce that in a plain environment and not doing imports in such
an directory is better anyway. so I will just leave this in as it is now.