Better error messages on module importing
When composing the app, and importing the app module, it loads dependencies, that, if absent, end up in an exception.
The exception message gets swallowed up, and no information about it reaches the debug log. Thus we have no idea what went wrong.
This commit displays the last entry of the stack trace of the inner exception.