Add a cli call to export config as json
In order for third party apps to be able to access the config, it needs to be in a standard format.
Go through each config value and unpack it from whatever data structure there is, and convert to a dictionary.

How deep to unpack is pretty arbitrary, but these considerations are followed:
* Usually when something is configured with a list/dict, we're interested in the values of the list/dict, so we unpack if that's the first thing we see.
* If it's a non-scalar, non-dict, non-list, try to get it as a __dict__ and end there, because this is usually the data structure (class) used for the config, and its members are either scalars or non-interesting complex members.
* If there is no __dict__ capability, just __repr__ it.

This strategy should cover the majority of instances where the config is read by a 3rd party script.